feat: Introduce police car spawning mechanism
Adds logic and state management for spawning police cars in the game. This includes updating level themes with new colors and initializing new state variables for police car spawns and warnings.

The police cars will now appear after a set delay and can be triggered based on game progression and timers. A visual warning will also be introduced to alert the player of incoming police.